requirements.txt
PyQt6
pandas
openpyxl
requests
python-docx
reportlab
